Move tchap funcs out of client.ts which is an element copied file
