+++ cypress/support/rooms.ts
@@ -5,14 +5,30 @@
 /// <reference types="cypress" />
 /// <reference types="@testing-library/cypress" />
 
-import TchapCreateRoom from "../../src/tchap/lib/createTchapRoom";
 import Chainable = Cypress.Chainable;
+import type { ICreateRoomOpts } from "matrix-js-sdk/src/@types/requests";
+import type { Room } from "matrix-js-sdk/src/models/room";
+
+import TchapCreateRoom from "../../src/tchap/lib/createTchapRoom";
 import { TchapRoomType } from "../../src/tchap/@types/tchap";
 
 declare global {
     // eslint-disable-next-line @typescript-eslint/no-namespace
     namespace Cypress {
         interface Chainable {
+            /**
+             * Leave a room.
+             * @param roomId the id of the room to invite to
+             */
+            leaveRoom(roomId: string): Chainable<{}>;
+
+            /**
+             * Leave a room. If the leaving fails, log and carry on without crashing the test.
+             * Needed until this is fixed : https://github.com/tchapgouv/synapse-manage-last-admin/issues/11
+             * @param roomId the id of the room to invite to
+             */
+            leaveRoomWithSilentFail(roomId: string): Chainable<{}>;
+
             /**
              * Leave the room that we are currently in.
              */
@@ -23,6 +39,13 @@ declare global {
              */
             leaveCurrentRoomWithSilentFail(): Chainable<{}>;
 
+            /**
+             * Create a room with given options.
+             * @param options the options to apply when creating the room
+             * @return the ID of the newly created room
+             */
+            createRoom(options: ICreateRoomOpts): Chainable<string>;
+
             createPublicRoom(roomName: string): Chainable<string>;
 
             createPrivateRoom(roomName: string): Chainable<string>;
@@ -31,6 +54,19 @@ declare global {
         }
     }
 }
+
+Cypress.Commands.add("leaveRoom", (roomId: string): Chainable<{}> => {
+    return cy.getClient().then((cli) => cli.leave(roomId));
+});
+
+Cypress.Commands.add("leaveRoomWithSilentFail", (roomId: string): Chainable<{}> => {
+    return cy.getClient().then((cli) => {
+        return cli.leave(roomId).catch((err) => {
+            cy.log("COULD NOT LEAVE ROOM ! Continuing silently.", err);
+            return {};
+        });
+    });
+});
 
 Cypress.Commands.add("leaveCurrentRoom", (): Chainable<{}> => {
     // We find the roomId to clean up from the current URL.
@@ -55,6 +91,28 @@ Cypress.Commands.add("leaveCurrentRoomWithSilentFail", (): Chainable<{}> => {
             console.error("Did not find roomId in url. Not cleaning up.");
             return {};
         }
+    });
+});
+
+Cypress.Commands.add("createRoom", (options: ICreateRoomOpts): Chainable<string> => {
+    return cy.window({ log: false }).then(async (win) => {
+        const cli = win.mxMatrixClientPeg.matrixClient;
+        const resp = await cli.createRoom(options);
+        const roomId = resp.room_id;
+
+        if (!cli.getRoom(roomId)) {
+            await new Promise<void>((resolve) => {
+                const onRoom = (room: Room) => {
+                    if (room.roomId === roomId) {
+                        cli.off(win.matrixcs.ClientEvent.Room, onRoom);
+                        resolve();
+                    }
+                };
+                cli.on(win.matrixcs.ClientEvent.Room, onRoom);
+            });
+        }
+
+        return roomId;
     });
 });
